reafctor: relocate components and imports
