+++ apps/web/src/app/(main)/(auth)/register/page.tsx
@@ -2,7 +2,7 @@
 
 import { registerAction } from "@/app/actions/register"
 import { useAuth } from "@/app/context/AuthContext"
-import ThirdPartyButtons from "@/components/layouts/Auth/ThirdPartyButtons"
+import { AuthThirdPartyProviders as ThirdPartyButtons } from "@/components/Auth"
 import { Button } from "@mav/ui/components/buttons"
 import { Form, type FormState, InputField } from "@mav/ui/components/fields"
 import Image from "next/image"
