+++ apps/web/src/app/(profile)/profile/[handle]/(user)/(characters)/characters/CharacterView.tsx
@@ -1,10 +1,10 @@
 "use client"
 
 import CreateFolderModal from "@/components/Modals/CreateFolder"
-import { CharacterCard } from "@/components/layouts/Cards"
-import FolderView from "@/components/layouts/Folders"
-import { SearchBox } from "@/components/layouts/Forms"
-import GridResponsive from "@/components/layouts/Layouts/GridResponsive"
+import { CharacterCard } from "@/components/Cards"
+import FolderView from "@/components/Folders"
+import { SearchBox } from "@/components/Forms"
+import { GridResponsive } from "@/components/Containers"
 import type { CharacterResponse, Folder } from "@/types/characters"
 import { folderColors } from "@/utils/constants"
 import { Button } from "@mav/ui/components/buttons"
