+++ apps/web/src/app/(profile)/profile/[handle]/(user)/(favorites)/favorites/loading.tsx
@@ -1,6 +1,5 @@
-import { FursonaCard } from "@/components/layouts/Cards"
-import GridResponsive from "@/components/layouts/Layouts/GridResponsive"
-import MarginClamp from "@/components/layouts/Layouts/MarginClamp"
+import { CharacterCard } from "@/components/Cards"
+import { MarginClamp, GridResponsive } from "@/components/Containers"
 
 export default function Loading() {
   return (
@@ -8,7 +7,7 @@ export default function Loading() {
       {/* For 8 times generate a dummy fursona card */}
       <GridResponsive breakpoint={250} className="gap-1.5" role="listbox">
         {Array.from({ length: 8 }, (_, index) => (
-          <FursonaCard loading={true} key={index} />
+          <CharacterCard loading={true} key={index} />
         ))}
       </GridResponsive>
     </MarginClamp>
