+++ apps/web/src/app/(profile)/profile/[handle]/(user)/(overview)/panels/InformationPanel.tsx
@@ -1,5 +1,5 @@
 import { type User } from "@/app/context/AuthContext"
-import Field from "@/components/layouts/Layouts/Field"
+import { Field } from "@/components/Containers"
 import type { UserType } from "@/types/users"
 import { Button } from "@mav/ui/components/buttons"
 import { Group } from "@mav/ui/components/layouts"
