+++ apps/web/src/components/Cards/ColorPalette.tsx
