+++ apps/web/src/components/Cards/PinnedCharacter.tsx
