+++ apps/web/src/components/Cards/index.tsx
@@ -2,3 +2,5 @@ export * from "./CharacterCard"
 export * from "./ColorPalette"
 export * from "./PinnedCharacter"
 export * from "./ReferenceCard"
+export * from "./PinnedCharacter"
+export * from "./Status"
