+++ apps/web/src/components/Modals/EditPanel.tsx
@@ -7,7 +7,7 @@ import Image from "next/image"
 import { useState } from "react"
 import { FaCode, FaComment, FaInfoCircle } from "react-icons/fa"
 import { LuXCircle } from "react-icons/lu"
-import Modal from "../layouts/Modal"
+import Modal from "./ModalBase"
 import Note from "../layouts/Note"
 
 export default function EditPanelModal({
