+++ apps/web/src/components/index.tsx
@@ -3,3 +3,4 @@ export * from "./CheckLocalSettings"
 export * from "./NoJSMessage"
 export * from "./SkipNav"
 export * from "./Analytics"
+export * from "./Separator"
